requirements.txt
nekos.py
discord.py
async_cse
matplotlib
sr_api
asyncdagpi
gtts
pyfiglet
aiosqlite
async_cleverbot
jishaku